+++ envoy-control-core/src/test/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/resource/listeners/filters/rbac/RBACFilterFactoryTest.kt
@@ -226,7 +226,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     fun `should generate RBAC rules for incoming permissions with log unlisted clients and endpoints`() {
         // given
         val policyName = "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
-            "clients=[ClientWithSelector(name=client1, selector=null), ClientWithSelector(name=client2, selector=null)], " +
+            "clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], " +
             "unlistedClientsPolicy=LOG, oauth=null)"
         val expectedShadowRules = expectedSimpleEndpointPermissionsJson(policyName)
 
@@ -241,7 +241,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                 "/example",
                     PathMatchingType.PATH,
                     setOf("GET", "POST"),
-                    setOf(ClientWithSelector("client1"), ClientWithSelector("client2")),
+                    setOf(ClientWithSelector.create("client1"), ClientWithSelector.create("client2")),
                     Incoming.UnlistedPolicy.LOG
                 )
             ),
@@ -259,7 +259,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     fun `should generate RBAC rules for incoming permissions with block unlisted endpoints and log clients`() {
         // given
         val policyName = "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
-            "clients=[ClientWithSelector(name=client1, selector=null), ClientWithSelector(name=client2, selector=null)], " +
+            "clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], " +
             "unlistedClientsPolicy=LOG, oauth=null)"
         val expectedShadowRules = expectedSimpleEndpointPermissionsJson(policyName)
 
@@ -274,7 +274,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                                 "/example",
                                 PathMatchingType.PATH,
                                 setOf("GET", "POST"),
-                                setOf(ClientWithSelector("client1"), ClientWithSelector("client2")),
+                                setOf(ClientWithSelector.create("client1"), ClientWithSelector.create("client2")),
                                 Incoming.UnlistedPolicy.LOG
                         )
                 ),
@@ -292,7 +292,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     fun `should generate RBAC rules for incoming permissions with log unlisted endpoints and block clients`() {
         // given
         val policyName = "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
-            "clients=[ClientWithSelector(name=client1, selector=null), ClientWithSelector(name=client2, selector=null)], " +
+            "clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], " +
             "unlistedClientsPolicy=BLOCKANDLOG, oauth=null)"
         val expectedShadowRules = expectedSimpleEndpointPermissionsJson(policyName)
 
@@ -307,7 +307,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                                 "/example",
                                 PathMatchingType.PATH,
                                 setOf("GET", "POST"),
-                                setOf(ClientWithSelector("client1"), ClientWithSelector("client2")),
+                                setOf(ClientWithSelector.create("client1"), ClientWithSelector.create("client2")),
                                 Incoming.UnlistedPolicy.BLOCKANDLOG
                         )
                 ),
@@ -325,7 +325,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     fun `should generate RBAC rules for incoming permissions with roles`() {
         // given
         val policyName = "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
-            "clients=[ClientWithSelector(name=role-1, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)"
+            "clients=[ClientWithSelector(name=role-1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)"
         val expectedRbacBuilder = getRBACFilter(expectedSimpleEndpointPermissionsJson(policyName))
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
@@ -333,8 +333,8 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
-                        setOf(ClientWithSelector("role-1"))
-                )), roles = listOf(Role("role-1", setOf(ClientWithSelector("client1"), ClientWithSelector("client2"))))
+                        setOf(ClientWithSelector.create("role-1"))
+                )), roles = listOf(Role("role-1", setOf(ClientWithSelector.create("client1"), ClientWithSelector.create("client2"))))
         )
 
         // when
@@ -355,17 +355,17 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
                         setOf(
-                                ClientWithSelector("client1"),
-                                ClientWithSelector("client1"),
-                                ClientWithSelector("client1", "selector"),
-                                ClientWithSelector("client1-duplicated", "selector"),
-                                ClientWithSelector("client1-duplicated"),
-                                ClientWithSelector("role-1")
+                                ClientWithSelector.create("client1"),
+                                ClientWithSelector.create("client1"),
+                                ClientWithSelector.create("client1", "selector"),
+                                ClientWithSelector.create("client1-duplicated", "selector"),
+                                ClientWithSelector.create("client1-duplicated"),
+                                ClientWithSelector.create("role-1")
                         )
                 )), roles = listOf(Role("role-1", setOf(
-                        ClientWithSelector("client1-duplicated"),
-                        ClientWithSelector("client1-duplicated"),
-                        ClientWithSelector("client2"))
+                        ClientWithSelector.create("client1-duplicated"),
+                        ClientWithSelector.create("client1-duplicated"),
+                        ClientWithSelector.create("client2"))
                 ))
         )
 
@@ -386,12 +386,12 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
-                        setOf(ClientWithSelector("client1"))
+                        setOf(ClientWithSelector.create("client1"))
                 ), IncomingEndpoint(
                         "/example2",
                         PathMatchingType.PATH,
                         setOf("POST"),
-                        setOf(ClientWithSelector("client2"))
+                        setOf(ClientWithSelector.create("client2"))
                 ))
         )
 
@@ -411,13 +411,13 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
-                        setOf(ClientWithSelector("role-1"))
+                        setOf(ClientWithSelector.create("role-1"))
                 ), IncomingEndpoint(
                         "/example2",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
-                        setOf(ClientWithSelector("client2"), ClientWithSelector("client1"))
-                )), roles = listOf(Role("role-1", setOf(ClientWithSelector("client1"), ClientWithSelector("client2"))))
+                        setOf(ClientWithSelector.create("client2"), ClientWithSelector.create("client1"))
+                )), roles = listOf(Role("role-1", setOf(ClientWithSelector.create("client1"), ClientWithSelector.create("client2"))))
         )
         val expectedRbacBuilder = getRBACFilter(expectedTwoClientsSimpleEndpointPermissionsJson(
             "${incomingPermission.endpoints[0]}", "${incomingPermission.endpoints[1]}"
@@ -439,13 +439,13 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
-                        setOf(ClientWithSelector("client2"), ClientWithSelector("role-1"))
+                        setOf(ClientWithSelector.create("client2"), ClientWithSelector.create("role-1"))
                 ), IncomingEndpoint(
                         "/example2",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
-                        setOf(ClientWithSelector("role-2"), ClientWithSelector("client1"))
-                )), roles = listOf(Role("role-1", setOf(ClientWithSelector("client1"))), Role("role-2", setOf(ClientWithSelector("client2"))))
+                        setOf(ClientWithSelector.create("role-2"), ClientWithSelector.create("client1"))
+                )), roles = listOf(Role("role-1", setOf(ClientWithSelector.create("client1"))), Role("role-2", setOf(ClientWithSelector.create("client2"))))
         )
         val expectedRbacBuilder = getRBACFilter(expectedTwoClientsSimpleEndpointPermissionsJson(
             "${incomingPermission.endpoints[0]}", "${incomingPermission.endpoints[1]}"
@@ -468,7 +468,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
-                        setOf(ClientWithSelector("client1"), ClientWithSelector("client2"))
+                        setOf(ClientWithSelector.create("client1"), ClientWithSelector.create("client2"))
                 ))
         )
 
@@ -604,7 +604,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
-                        setOf(ClientWithSelector("client1"))
+                        setOf(ClientWithSelector.create("client1"))
                 ))
         )
 
@@ -625,7 +625,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
-                        setOf(ClientWithSelector("client1"), ClientWithSelector("client2"))
+                        setOf(ClientWithSelector.create("client1"), ClientWithSelector.create("client2"))
                 ))
         )
 
@@ -649,7 +649,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
-                        setOf(ClientWithSelector("client2", "selector"))
+                        setOf(ClientWithSelector.create("client2", "selector"))
                 ))
         )
 
@@ -673,7 +673,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
-                        setOf(ClientWithSelector("client1", "selector"))
+                        setOf(ClientWithSelector.create("client1", "selector"))
                 ))
         )
 
@@ -697,11 +697,11 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
-                        setOf(ClientWithSelector("role1"))
+                        setOf(ClientWithSelector.create("role1"))
                 )),
                 roles = listOf(Role("role1", setOf(
-                        ClientWithSelector("client1", "selector1"),
-                        ClientWithSelector("client2", "selector2"))
+                        ClientWithSelector.create("client1", "selector1"),
+                        ClientWithSelector.create("client2", "selector2"))
                 ))
         )
 
@@ -725,7 +725,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                 "/example",
                 PathMatchingType.PATH,
                 setOf("GET"),
-                setOf(ClientWithSelector("client1"))
+                setOf(ClientWithSelector.create("client1"))
             ))
         )
 
@@ -788,7 +788,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedEndpointPermissionsWithDifferentRulesForDifferentClientsJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -808,7 +808,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                 ${originalAndAuthenticatedPrincipal("client1")}
               ]
             },
-            "IncomingEndpoint(path=/example2, pathMatchingType=PATH, methods=[POST], clients=[ClientWithSelector(name=client2, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example2, pathMatchingType=PATH, methods=[POST], clients=[ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -835,7 +835,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpFromDiscoveryWithSelectorAuthPermissionsJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=selector)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=selector, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -870,7 +870,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpWithSelectorAuthPermissionsJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client2, selector=selector)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client2, selector=selector, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -905,7 +905,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpWithStaticRangeAndSelectorAuthPermissionsAndRolesJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=role1, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=role1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -949,7 +949,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpAuthPermissionsJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null), ClientWithSelector(name=client2, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -1008,7 +1008,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         {
           "policies": {
            """ /* notice that duplicated clients occurs only once here */ + """
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null), ClientWithSelector(name=client1, selector=selector), ClientWithSelector(name=client1-duplicated, selector=selector), ClientWithSelector(name=client1-duplicated, selector=null), ClientWithSelector(name=role-1, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client1, selector=selector, negated=false), ClientWithSelector(name=client1-duplicated, selector=selector, negated=false), ClientWithSelector(name=client1-duplicated, selector=null, negated=false), ClientWithSelector(name=role-1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -1228,7 +1228,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpAuthWithStaticRangeJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -1255,7 +1255,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpAuthWithStaticRangeAndSourceIpJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null), ClientWithSelector(name=client2, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -1292,7 +1292,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedEndpointPermissionsLogUnlistedEndpointsAndBlockUnlistedClients = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null), ClientWithSelector(name=client2, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                     "and_rules": {
@@ -1348,7 +1348,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private fun expectedPoliciesForAllowedClient(principals: String) = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {