+++ CHANGELOG.md
@@ -10,10 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [0.19.25]
 
 ### Changed
-
+- Prefix for negating values from jwt token used in rbac
 - Configurable default clients lists
-
-### Changed
 
 ## [0.19.24]